% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDECquery.R
\name{CDECquery}
\alias{CDECquery}
\title{CDECquery}
\usage{
CDECquery(id, sensor, interval = "D", start, end)
}
\arguments{
\item{id}{The three letter station ID, quoted and in all caps (for a full list of station ids, see: )}

\item{sensor}{The number of the sensor you want (for a full list of sensors, see: )}

\item{interval}{The code of the data duration you want: "D" for daily (default), "E" for event, or "M" monthly.}

\item{start}{The start date of the data, in mm/dd/yyyy format}

\item{end}{The end date of the data you want, in mm/dd/yyyy format}
}
\value{
The data between the specified \code{start_date} and \code{end_date}, in table format
}
\description{
CDECquery function, recreated from the sharpshootR package (function author: Dylan Beaudette)
}
\details{
For original documentation, see ?sharpshootR::CDECquery.  Dates must be in yyyy-mm-dd format, type of data and station is quoted, sensor number is numeric.
}
\examples{
d <- CDECquery("FPT", 20, "E", "03/01/2011", "05/01/2011")
head(d)
}
