% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWR.R
\docType{package}
\name{DWR}
\alias{DWR}
\alias{DWR-package}
\title{DWR: a package for getting data from cdec.water.ca.gov}
\description{
The DWR package contains a single function (CDECquery) and two built-in datasets (sensorIDs and stationIDs).
}
\section{CDECquery function}{

This function is authored by Dylan Beaudette and implemented here as a standalone-version of the one in the sharpshootR package (https://CRAN.R-project.org/package=sharpshootR).  Documentation has been generated to include the "E" duration type, as it's worked for us so far without modifying the original function (although limited testing has been done).
}

