% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCDEC.R, R/getCDEC_recreated.R
\docType{data}
\name{f2}
\alias{f2}
\alias{f2}
\title{getCDEC - a function generated by getXML::createFunction}
\format{An object of class \code{list} of length 7.}
\usage{
f2

f2
}
\arguments{
\item{sensor_num}{The number of the sensor you want (for a full list of sensors, see: )}

\item{start_date}{The start date of the data, in mm/dd/yyyy format}

\item{end_date}{The end date of the data you want, in mm/dd/yyyy format}

\item{station_id}{The three letter station ID, quoted and in all caps (for a full list of station ids, see: )}

\item{dur_code}{The code of the data duration you want.}

\item{sensor_num}{The number of the sensor you want (for a full list of sensors, see: )}

\item{start_date}{The start date of the data, in mm/dd/yyyy format}

\item{end_date}{The end date of the data you want, in mm/dd/yyyy format}

\item{station_id}{The three letter station ID, quoted and in all caps (for a full list of station ids, see: )}

\item{dur_code}{The code of the data duration you want.}
}
\value{
The data between the specified \code{start_date} and \code{end_date}, in table format

The data between the specified \code{start_date} and \code{end_date}, in table format
}
\description{
getCDEC - a function generated by getXML::createFunction

getCDEC - a function generated by getXML::createFunction
}
\details{
dates must be in mm/dd/yyyy format, type of data and station is quoted, sensor number is numeric.

dates must be in mm/dd/yyyy format, type of data and station is quoted, sensor number is numeric.
}
\examples{
d <- getCDEC(20, "03/01/2011", "05/01/2011", "FPT", "E")
head(d)
d <- getCDEC(20, "03/01/2011", "05/01/2011", "FPT", "E")
head(d)
}
\keyword{datasets}
