% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorIDs.R
\docType{data}
\name{sensorIDs}
\alias{sensorIDs}
\title{Sensor IDs of CDEC}
\format{A dataframe with 243 observations of 5 variables:
\describe{
\item{Sensor No.}{Numeric sensor number, which is one of the required arguments for the getCDEC function}
\item{Sensor}{Abbreviated description of the sensor and what it records}
\item{PE Code}{No idea what this is or what it's for}
\item{Description}{Longer description of the sensor and what it records}
\item{Units}{Units that the sensor records in}
}}
\source{
\url{http://cdec.water.ca.gov/misc/senslist.html}
}
\usage{
sensorIDs
}
\description{
A dataframe containing the numeric sensor code and descriptions for the water quality stations from cdec.water.ca.gov.
}
\keyword{datasets}
