% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationIDs.R
\docType{data}
\name{stationIDs}
\alias{stationIDs}
\title{Station IDs of CDEC}
\format{A dataframe with 2605 observations of 2 variables:
\describe{
\item{StationID}{Abbreviated station ID, which is one of the required arguments for the getCDEC function}
\item{Description}{Brief description of the station, scraped from cdec.water.ca.gov}
}}
\usage{
stationIDs
}
\description{
A dataframe containing the 3-letter station code and the description of water quality stations from cdec.water.ca.gov/.
}
\keyword{datasets}
