% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDECquery.R
\name{CDECquery}
\alias{CDECquery}
\title{CDECquery}
\usage{
CDECquery(id, sensor, interval = "D", start, end)
}
\arguments{
\item{id}{The three letter station ID, quoted and in all caps (for a full list of station ids, see: )}

\item{sensor}{The number of the sensor you want (for a full list of sensors, see: )}

\item{interval}{The code of the data duration you want: "D" for daily (default), "E" for event, or "M" monthly.}

\item{start}{The start date of the data, in yyyy-mm-dd format}

\item{end}{The end date of the data you want, in yyyy-mm-dd format}
}
\value{
The data between the specified \code{start_date} and \code{end_date}, in table format
}
\description{
CDECquery function, edited from the sharpshootR package (original function author: Dylan Beaudette)
}
\details{
For original documentation, see ?sharpshootR::CDECquery.  For argument inputs, start and end dates must be in yyyy-mm-dd format, duration code and station id are quoted, sensor number is numeric.
}
